Fix transport error on delete session

+++ src/Ydb.Sdk/src/Services/Table/Session.cs
@@ -12,13 +12,16 @@ public partial class Session : ClientBase, IDisposable
     private readonly ILogger _logger;
     private bool _disposed;
 
-    internal Session(Driver driver, SessionPool? sessionPool, string id, string? endpoint)
+    private readonly bool _deleteOnDispose;
+
+    internal Session(Driver driver, SessionPool? sessionPool, string id, string? endpoint, bool deleteOnDispose = true)
         : base(driver)
     {
         _sessionPool = sessionPool;
         _logger = Driver.LoggerFactory.CreateLogger<Session>();
         Id = id;
         Endpoint = endpoint;
+        _deleteOnDispose = deleteOnDispose;
     }
 
     public string Id { get; }
@@ -73,13 +76,17 @@ public partial class Session : ClientBase, IDisposable
         {
             if (_sessionPool is null)
             {
-                _logger.LogTrace($"Closing detached session on dispose: {Id}");
-
-                var client = new TableClient(Driver, new NoPool());
-                _ = client.DeleteSession(Id, new DeleteSessionSettings
+                if (_deleteOnDispose)
                 {
-                    TransportTimeout = DeleteSessionTimeout
-                });
+                    _logger.LogTrace($"Closing detached session on dispose: {Id}");
+
+                    var client = new TableClient(Driver, new NoPool());
+                    var task = client.DeleteSession(Id, new DeleteSessionSettings
+                    {
+                        TransportTimeout = DeleteSessionTimeout
+                    });
+                    task.Wait();
+                }
             }
             else
             {
